Add some documentation to the raycast functions.
Since these are now functions and not in a class there's no obvious
encapsulation for the two related functions. Previously they were
only available in the PolyVox namespace documentation where they were
hidden among lots of other things. Adding the \file command to
Raycast.h allows up to add comments to that page (_raycast_8h.html)
and have that as the central location for the raycast docs.

There still needs to be a comment about the form that the callback
must take (return type and arguments) including the fact that both
functors and lambdas work.