On systems that support it, raise SIGTRAP to drop into the debugger
Clang was being clever and giving a warning for this line so it prompted me
to fix it. I believe that SIGTRAP is the correct way to get the debugger to
work here.

It does a compile-time check for the platform when not using MSVC.

Discussed at http://www.volumesoffun.com/phpBB3/viewtopic.php?p=3766#p3766